# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Migration Factory API",
  "version": "1.0.0"
 },
 "paths": {
  "/api/v1/projects": {
   "post": {
    "tags": [
     "projects"
    ],
    "summary": "Create Project",
    "operationId": "create_project_api_v1_projects_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "projects"
    ],
    "summary": "List Projects",
    "operationId": "list_projects_api_v1_projects_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ProjectOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/projects/{project_id}/members": {
   "get": {
    "tags": [
     "projects"
    ],
    "summary": "List Members",
    "operationId": "list_members_api_v1_projects__project_id__members_get",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/MemberOut"
         },
         "title": "Response List Members Api V1 Projects  Project Id  Members Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "projects"
    ],
    "summary": "Put Member",
    "operationId": "put_member_api_v1_projects__project_id__members_put",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MemberIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MemberOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/projects/{project_id}/scans": {
   "post": {
    "tags": [
     "projects"
    ],
    "summary": "Create Scan",
    "operationId": "create_scan_api_v1_projects__project_id__scans_post",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     },
     {
      "name": "idempotency-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ScanCreate"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScanAccepted"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "projects"
    ],
    "summary": "List Scans",
    "operationId": "list_scans_api_v1_projects__project_id__scans_get",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 20,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ScanOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/jobs/{job_id}": {
   "get": {
    "tags": [
     "executions"
    ],
    "summary": "Get Job",
    "operationId": "get_job_api_v1_jobs__job_id__get",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/auth/dev-login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Dev Login",
    "operationId": "dev_login_api_v1_auth_dev_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DevLogin"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/profiles": {
   "get": {
    "tags": [
     "profiles"
    ],
    "summary": "List Profiles",
    "operationId": "list_profiles_api_v1_profiles_get",
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(draft|validated|retired)$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "runtime",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(spring|quarkus)$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Runtime"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ProfileOut"
         },
         "title": "Response List Profiles Api V1 Profiles Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "profiles"
    ],
    "summary": "Create Profile",
    "description": "New immutable profile version (draft). Existing versions are never edited.",
    "operationId": "create_profile_api_v1_profiles_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProfileCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProfileOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/uploads": {
   "post": {
    "tags": [
     "sources"
    ],
    "summary": "Upload Zip",
    "description": "Body = the ZIP bytes (Content-Type: application/zip). Size enforced while streaming, structure validated before storing.",
    "operationId": "upload_zip_api_v1_uploads_post",
    "parameters": [
     {
      "name": "x-filename",
      "in": "header",
      "required": false,
      "schema": {
       "type": "string",
       "maxLength": 255,
       "default": "project.zip",
       "title": "X-Filename"
      }
     },
     {
      "name": "content-length",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Content-Length"
      }
     }
    ],
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UploadOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "DevLogin": {
    "properties": {
     "username": {
      "type": "string",
      "title": "Username"
     }
    },
    "type": "object",
    "required": [
     "username"
    ],
    "title": "DevLogin"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "JobOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "state": {
      "type": "string",
      "title": "State"
     },
     "phase": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phase"
     },
     "progress": {
      "type": "integer",
      "title": "Progress"
     },
     "attempts": {
      "type": "integer",
      "title": "Attempts"
     },
     "cancelRequested": {
      "type": "boolean",
      "title": "Cancelrequested"
     },
     "errorCode": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Errorcode"
     },
     "errorMessage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Errormessage"
     },
     "createdBy": {
      "type": "string",
      "title": "Createdby"
     },
     "createdAt": {
      "type": "string",
      "format": "date-time",
      "title": "Createdat"
     },
     "startedAt": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Startedat"
     },
     "finishedAt": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finishedat"
     }
    },
    "type": "object",
    "required": [
     "id",
     "kind",
     "state",
     "phase",
     "progress",
     "attempts",
     "cancelRequested",
     "errorCode",
     "errorMessage",
     "createdBy",
     "createdAt",
     "startedAt",
     "finishedAt"
    ],
    "title": "JobOut"
   },
...
   "MemberIn": {
    "properties": {
     "subjectId": {
      "type": "string",
      "maxLength": 255,
      "minLength": 1,
      "title": "Subjectid"
     },
     "role": {
      "type": "string",
      "enum": [
       "owner",
       "architect",
       "developer",
       "reviewer",
       "auditor"
      ],
      "title": "Role"
     }
    },
    "type": "object",
    "required": [
     "subjectId",
     "role"
    ],
    "title": "MemberIn"
   },
   "MemberOut": {
    "properties": {
     "subjectId": {
      "type": "string",
      "title": "Subjectid"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "displayName": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Displayname"
     }
    },
    "type": "object",
    "required": [
     "subjectId",
     "role"
    ],
    "title": "MemberOut"
   },
...
   "Page_ProjectOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ProjectOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "nextCursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Nextcursor"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total"
    ],
    "title": "Page[ProjectOut]"
   },
...
   "Page_ScanOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ScanOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "nextCursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Nextcursor"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total"
    ],
    "title": "Page[ScanOut]"
   },
...
   "ProfileOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "key": {
      "type": "string",
      "title": "Key"
     },
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "runtime": {
      "type": "string",
      "title": "Runtime"
     },
     "javaVersion": {
      "type": "string",
      "title": "Javaversion"
     },
     "camelVersion": {
      "type": "string",
      "title": "Camelversion"
     },
     "runtimeVersion": {
      "type": "string",
      "title": "Runtimeversion"
     },
     "bomCoordinates": {
      "items": {},
      "type": "array",
      "title": "Bomcoordinates"
     },
     "tooling": {
      "additionalProperties": true,
      "type": "object",
      "title": "Tooling"
     },
     "dependencyMapping": {
      "additionalProperties": true,
      "type": "object",
      "title": "Dependencymapping"
     },
     "sources": {
      "items": {},
      "type": "array",
      "title": "Sources"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "digest": {
      "type": "string",
      "title": "Digest"
     },
     "createdAt": {
      "type": "string",
      "format": "date-time",
      "title": "Createdat"
     }
    },
    "type": "object",
    "required": [
     "id",
     "key",
     "version",
     "name",
     "runtime",
     "javaVersion",
     "camelVersion",
     "runtimeVersion",
     "bomCoordinates",
     "tooling",
     "dependencyMapping",
     "sources",
     "status",
     "digest",
     "createdAt"
    ],
    "title": "ProfileOut"
   },
   "ProjectCreate": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 2,
      "pattern": "^[\\w .\\-()]+$",
      "title": "Name"
     },
     "description": {
      "type": "string",
      "maxLength": 2000,
      "title": "Description",
      "default": ""
     },
     "sourceType": {
      "type": "string",
      "enum": [
       "local",
       "zip",
       "git"
      ],
      "title": "Sourcetype"
     },
     "targetRuntime": {
      "type": "string",
      "enum": [
       "spring",
       "quarkus"
      ],
      "title": "Targetruntime"
     },
     "repositoryUrl": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 1000
       },
       {
        "type": "null"
       }
      ],
      "title": "Repositoryurl"
     },
     "defaultBranch": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^[\\w./-]{1,200}$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Defaultbranch"
     },
     "credentialRef": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Credentialref"
     },
     "localPath": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 1000
       },
       {
        "type": "null"
       }
      ],
      "title": "Localpath"
     },
     "uploadId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Uploadid"
     }
    },
    "type": "object",
    "required": [
     "name",
     "sourceType",
     "targetRuntime"
    ],
    "title": "ProjectCreate"
   },
   "ProjectOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "sourceType": {
      "type": "string",
      "title": "Sourcetype"
     },
     "repositoryUrl": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Repositoryurl"
     },
     "repositoryProvider": {
      "type": "string",
      "title": "Repositoryprovider"
     },
     "credentialRef": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Credentialref"
     },
     "localPath": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Localpath"
     },
     "upload": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Upload"
     },
     "defaultBranch": {
      "type": "string",
      "title": "Defaultbranch"
     },
     "targetRuntime": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Targetruntime"
     },
     "requiredGates": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Requiredgates"
     },
     "independentReview": {
      "type": "boolean",
      "title": "Independentreview"
     },
     "prConfig": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Prconfig"
     },
     "myRole": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Myrole"
     },
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "createdAt": {
      "type": "string",
      "format": "date-time",
      "title": "Createdat"
     },
     "lastScan": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lastscan"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "description",
     "sourceType",
     "repositoryUrl",
     "repositoryProvider",
     "credentialRef",
     "defaultBranch",
     "targetRuntime",
     "requiredGates",
     "independentReview",
     "prConfig",
     "myRole",
     "version",
     "createdAt"
    ],
    "title": "ProjectOut"
   },
...
   "ScanAccepted": {
    "properties": {
     "scanId": {
      "type": "string",
      "format": "uuid",
      "title": "Scanid"
     },
     "jobId": {
      "type": "string",
      "format": "uuid",
      "title": "Jobid"
     },
     "status": {
      "type": "string",
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "scanId",
     "jobId",
     "status"
    ],
    "title": "ScanAccepted"
   },
   "ScanCreate": {
    "properties": {
     "ref": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^[\\w./-]{1,200}$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ref"
     },
     "catalogId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Catalogid"
     },
     "target": {
      "type": "string",
      "enum": [
       "spring",
       "quarkus"
      ],
      "title": "Target",
      "default": "spring"
     },
     "projectRoot": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500,
        "pattern": "^[\\w.\\-]+(/[\\w.\\-]+)*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Projectroot"
     }
    },
    "type": "object",
    "title": "ScanCreate"
   },
...
  }
 }
}